Move more transformers, rename GooseRest to Ferns
