+++ src/api.test.ts
@@ -5,7 +5,7 @@ import mongoose from "mongoose";
 import qs from "qs";
 import supertest from "supertest";
 
-import {gooseRestRouter} from "./api";
+import {fernsRouter} from "./api";
 import {setupAuth} from "./auth";
 import {Permissions} from "./permissions";
 import {
@@ -41,7 +41,7 @@ describe("ferns-api", () => {
       let deleteCalled = false;
       app.use(
         "/food",
-        gooseRestRouter(FoodModel, {
+        fernsRouter(FoodModel, {
           permissions: {
             list: [Permissions.IsAny],
             create: [Permissions.IsAny],
@@ -109,7 +109,7 @@ describe("ferns-api", () => {
 
       app.use(
         "/food",
-        gooseRestRouter(FoodModel, {
+        fernsRouter(FoodModel, {
           permissions: {
             list: [Permissions.IsAny],
             create: [Permissions.IsAny],
@@ -148,7 +148,7 @@ describe("ferns-api", () => {
       let deleteCalled = false;
       app.use(
         "/food",
-        gooseRestRouter(FoodModel, {
+        fernsRouter(FoodModel, {
           permissions: {
             list: [Permissions.IsAny],
             create: [Permissions.IsAny],
@@ -252,7 +252,7 @@ describe("ferns-api", () => {
       setupAuth(app, UserModel as any);
       app.use(
         "/food",
-        gooseRestRouter(FoodModel, {
+        fernsRouter(FoodModel, {
           permissions: {
             list: [Permissions.IsAdmin],
             create: [Permissions.IsAdmin],
@@ -400,7 +400,7 @@ describe("ferns-api", () => {
       setupAuth(app, UserModel as any);
       app.use(
         "/food",
-        gooseRestRouter(FoodModel, {
+        fernsRouter(FoodModel, {
           permissions: {
             list: [Permissions.IsAny],
             create: [Permissions.IsAuthenticated],
@@ -568,7 +568,7 @@ describe("ferns-api", () => {
       setupAuth(app, UserModel as any);
       app.use(
         "/users",
-        gooseRestRouter(UserModel, {
+        fernsRouter(UserModel, {
           permissions: {
             list: [Permissions.IsAuthenticated],
             create: [Permissions.IsAuthenticated],
