+++ src/auth.test.ts
@@ -2,10 +2,11 @@ import {assert} from "chai";
 import express from "express";
 import supertest from "supertest";
 
-import {AdminOwnerTransformer, gooseRestRouter} from "./api";
+import {fernsRouter} from "./api";
 import {setupAuth} from "./auth";
 import {Permissions} from "./permissions";
 import {Food, FoodModel, getBaseServer, setupDb, UserModel} from "./tests";
+import {AdminOwnerTransformer} from "./transformers";
 
 describe("auth tests", function () {
   let app: express.Application;
@@ -39,7 +40,7 @@ describe("auth tests", function () {
     setupAuth(app, UserModel as any);
     app.use(
       "/food",
-      gooseRestRouter(FoodModel, {
+      fernsRouter(FoodModel, {
         permissions: {
           list: [Permissions.IsAny],
           create: [Permissions.IsAuthenticated],
